[refactor] auth: 관리자 계정 자동 생성 관련 user -> auth로 도메인 이동
