+++ src/main/java/org/example/ootoutfitoftoday/domain/user/service/command/UserCommandServiceImpl.java
@@ -21,29 +21,4 @@ public class UserCommandServiceImpl implements UserCommandService {
     public void save(User user) {
         userRepository.save(user);
     }
-
-    @Override
-    public void initializeAdmin(
-            String loginId,
-            String email,
-            String nickname,
-            String username,
-            String password,
-            String phoneNumber
-    ) {
-        if (!userQueryService.existsByLoginId(loginId)) {
-            User admin = User.createAdmin(
-                    loginId,
-                    email,
-                    nickname,
-                    username,
-                    passwordEncoder.encode(password),
-                    phoneNumber
-            );
-            save(admin);
-            System.out.println("관리자 계정 초기 생성 완료되었습니다.");
-        } else {
-            System.out.println("관리자 계정이 이미 존재합니다.");
-        }
-    }
 }
